% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_curve_graph.R
\name{age_curve_graph}
\alias{age_curve_graph}
\title{Create a graph for aging curves, based on particular stats}
\usage{
age_curve_graph(stat, playertype = "batter", startyear = 2000,
  endyear = 2016, save = FALSE, path = getwd())
}
\arguments{
\item{stat}{A metric available through the FanGraphs database.}

\item{startyear}{The first year from which to calculate the aging curve. Defaults to 2000.}

\item{endyear}{The final year from which to calculate the aging curve. Defaults to 2016.}

\item{save}{Whether or not to save the graph. Defaults to FALSE.}

\item{path}{Where to save the graph. Defaults to the current working directory}

\item{playertyoe}{Whether to run the aging curves for batters or pitchers. Defaults to "batter"}
}
\description{
Create a graph for aging curves, based on particular stats
}
\examples{
\dontrun{age_curve_graph(stat, playertype = "batter", startyear = 2000, endyear = 2016, save = FALSE, path = getwd())}
}

