% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcher_graph.R
\name{catcher_graph}
\alias{catcher_graph}
\title{Graph called strike zone map by catcher}
\usage{
catcher_graph(catcher = "all", startdate = "2016-01-01",
  enddate = "2016-12-31", year = 0, count = "all", stand = "R\\",\\"S",
  save = FALSE, path = getwd())
}
\arguments{
\item{startdate}{Start date. Defaults to beginning of 2015. Format yyyy-mm-dd.}

\item{enddate}{End date. Defaults to end of 2015. Format yyyy-mm-dd.}

\item{year}{The year. Overrides startdate and enddate. Can only be one full year -- use startdate and enddate for full control over date range.}

\item{count}{Ball-strike count. Defaults to all. Input x instead of a number to get all counts; e.g. "3-x" for all 3-ball counts}

\item{stand}{Batter handedness. Defaults to both.}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}

\item{umpire}{Umpire name or ID. Defaults to "all". WARNING: Leaving as all will make graph take a long time to create.}
}
\description{
Graph called strike zone map by catcher
}

