% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datify.R
\name{datify}
\alias{datify}
\title{Turn two dates into a clean word description of the time}
\usage{
datify(startdate, enddate)
}
\arguments{
\item{startdate}{First date}

\item{enddate}{Last date}
}
\description{
Turn two dates into a clean word description of the time
}
\examples{
\dontrun{datify(startdate, enddate)}
}

