% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Extend a day or month number into a 2-character string}
\usage{
extend(n)
}
\arguments{
\item{n}{character string?}
}
\description{
Extend a day or month number into a 2-character string
}
\examples{
\dontrun{extend(n)}
}

