% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGQuery.R
\name{FGQuery}
\alias{FGQuery}
\title{Send a query to the FanGraphs database}
\usage{
FGQuery(query)
}
\arguments{
\item{query}{A query written in a way that is compatible with MySQL syntax}
}
\description{
Send a query to the FanGraphs database
}
\examples{
\dontrun{FGQuery(query)}
}

