% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_graph.R
\name{hist_graph}
\alias{hist_graph}
\title{Graph a histogram of all players for a given stat; highlight selected player.}
\usage{
hist_graph(player, stat, playertype = "batter", year = 2016, denom = "PA",
  mindenom = 200, bins = 10, save = FALSE, path = getwd())
}
\arguments{
\item{player}{The player to examine.}

\item{stat}{The stat to examine.}

\item{playertype}{Either "batter" or "pitcher". Defaults to batter.}

\item{year}{The year to graph.}

\item{denom}{The type of denominator to set the minimum. Defaults to plate appearances/total batters faced depending on the player type. Other options are "Pitches", "AB" (batters), "BIPcount", "GB", "FB", "LD"}

\item{mindenom}{The minimum number of whatever the denominator is to create the set of players to compare to. Default is 200 (plate appearances).}

\item{bins}{The number of bins on the histogram.}

\item{save}{Whether to save the graph. Defaults to FALSE.}

\item{path}{Where to save the graph. Defaults to the current working directory.}
}
\description{
Graph a histogram of all players for a given stat; highlight selected player.
}

