% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fg_db.R
\name{init_fg_db}
\alias{init_fg_db}
\title{Initialize a username and password for the FanGraphs database}
\usage{
init_fg_db()
}
\description{
Initialize a username and password for the FanGraphs database
}
\examples{
\dontrun{FGQuery(query)}
}

