% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_graph.R
\name{lineup_graph}
\alias{lineup_graph}
\title{Graph projected wRC+ for lineups in a game}
\usage{
lineup_graph(year, month, day, hometeam, awayteam, dh = 1, save = FALSE,
  path = getwd(), title = "")
}
\arguments{
\item{year}{Four-digit year}

\item{month}{Month}

\item{day}{Day}

\item{hometeam}{id for the home team}

\item{awayteam}{id for the away team}

\item{dh}{Whether it was a doubleheader. Defaults to 1.}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}

\item{title}{Title to place on the graph. No title by default}
}
\description{
Graph projected wRC+ for lineups in a game
}
\examples{
\dontrun{lineup_graph(2016, 04, 07, 12, 11, dh = 1, save = FALSE, path = getwd(), title = "")}
}

