% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfx_graph.R
\name{pfx_graph}
\alias{pfx_graph}
\title{Graph pitchf/x data -- 2 variables on different axes}
\usage{
pfx_graph(pitcher, x, y, startdate = "2016-01-01", enddate = "2016-12-31",
  pitchtypes = "all", stand = "R','L", color.code = TRUE, save = FALSE,
  path = getwd())
}
\arguments{
\item{pitcher}{Pitcher name or FanGraphs playerid.}

\item{x}{The variable to go on the x-axis.}

\item{y}{The variable to go on the y-axis.}

\item{startdate}{Start date. Defaults to beginning of 2016. Format yyyy-mm-dd.}

\item{enddate}{End date. Defaults to end of 2016. Format yyyy-mm-dd.}

\item{pitchtypes}{Defaults to all. To include multiple pitch types, create an array: c("Four-seam fastball","Changeup","Breaking ball")}

\item{stand}{Batter handedness. Defaults to both.}

\item{color.code}{Whether or not to color code the pitches by pitch type. Defaults to TRUE.}

\item{save}{Whether to save the graph. Defaults to FALSE.}

\item{path}{Where to save the graph. Defaults to the current working directory.}
}
\description{
Graph pitchf/x data -- 2 variables on different axes
}
\examples{
\dontrun{pfx_graph()}
}

