% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_date.R
\name{redo_date}
\alias{redo_date}
\title{Reorder a date so it's not in dumb yyyy-mm-dd format}
\usage{
redo_date(date)
}
\arguments{
\item{date}{A date to be passed to the function for reordering}
}
\description{
Reorder a date so it's not in dumb yyyy-mm-dd format
}
\examples{
\dontrun{redo_date(2016-05-02)}
}

