% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_graph.R
\name{release_graph}
\alias{release_graph}
\title{Graph pitcher release points}
\usage{
release_graph(pitcher, startdate = "2016-01-01", enddate = "2016-12-31",
  pitchtypes = "all", stand = "R','L", save = FALSE, path = getwd())
}
\arguments{
\item{pitcher}{Pitcher id}

\item{startdate}{Start date}

\item{enddate}{End date}

\item{pitchtypes}{Defaults to "all"}

\item{stand}{Batter handedness.}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}
}
\description{
Graph pitcher release points
}
\examples{
\dontrun{release_graph()}
}

