% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team}
\alias{team}
\title{Something about team :)}
\usage{
team(x, hometeam, awayteam)
}
\arguments{
\item{x}{Not sure}

\item{hometeam}{ID for the hometeam}

\item{awayteam}{ID for the awayteam}
}
\description{
Something about team :)
}

