% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umpire_graph.R
\name{umpire_graph}
\alias{umpire_graph}
\title{Graph pitches by umpire}
\usage{
umpire_graph(umpire = "all", startdate = "2015-01-01",
  enddate = "2015-12-31", count = "all", stand = "R\\",\\"S",
  save = FALSE, path = getwd())
}
\arguments{
\item{umpire}{Umpire ID. Defaults to "all".}

\item{startdate}{Start date}

\item{enddate}{End date}

\item{count}{Ball-strike count. Defaults to all}

\item{stand}{Batter handedness}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}
}
\description{
Graph pitches by umpire
}

