% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yty_graph.R
\name{yty_graph}
\alias{yty_graph}
\title{Graph the year-to-year correcltions between metrics}
\usage{
yty_graph(stat1, stat2 = stat1, playertype = "batter", startyear = 2000,
  endyear = 2016, minpa = 0, year.difference = 1, save = FALSE,
  path = getwd())
}
\arguments{
\item{stat1}{First statistic to correlate}

\item{stat2}{Second statistic to correlate}

\item{playertype}{Either "batter" or "pitcher". Defaults to "batter".}

\item{startyear}{Defaults to 2000}

\item{endyear}{Defaults to 2016}

\item{minpa}{Minimum number of plate appearances for a player's season to be included. Defaults to 0.}

\item{year.difference}{Defaults to 1}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}
}
\description{
Graph the year-to-year correcltions between metrics
}

