% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_graph.R
\name{lineup_graph}
\alias{lineup_graph}
\title{Graph projected wRC+ for lineups in a game}
\usage{
lineup_graph(year, month, day, hometeam, awayteam, dh = 1, title = NA,
  save = FALSE, path = getwd())
}
\arguments{
\item{year}{The year of the game.}

\item{month}{The month of the game.}

\item{day}{The day of the game.}

\item{hometeam}{Name or abbrevation for the home team.}

\item{awayteam}{Name or abbrevation for the away team}

\item{dh}{Doubleheader. Defaults to 1. 1 means either no doubleheader or the first game of a doubleheader; 2 means the second game.}

\item{title}{Title to place on the graph. Defaults to automatic titling.}

\item{save}{Whether to save the graph. Defaults to FALSE.}

\item{path}{Where to save the graph. Defaults to the current working directory.}
}
\description{
Graph projected wRC+ for lineups in a game
}
\examples{
\dontrun{lineup_graph(2016, 04, 07, 12, 11, dh = 1, save = FALSE, path = getwd(), title = "")}
}

