% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_graph.R
\name{percentile_graph}
\alias{percentile_graph}
\title{Graph percentiles}
\usage{
percentile_graph(player, player2 = "", playertype = "batter", year = 2016,
  stats = "standard", denom = "PA", mindenom = 200, save = FALSE,
  path = getwd())
}
\arguments{
\item{player}{XXXXX}
}
\description{
Graph percentiles
}

