% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sz_graph.R
\name{sz_graph}
\alias{sz_graph}
\title{Graph sz}
\usage{
sz_graph(batter = "all", pitcher = "all", count = "all",
  startdate = "2016-01-01", enddate = "2016-12-31", stand = "R','L",
  throws = "R','L", size = 3, heatmap = FALSE, pitchtypes = "all",
  color.code = "pitchtype", save = FALSE, path = getwd())
}
\arguments{
\item{batter}{Defaults to "all"}

\item{pitcher}{Defaults to "all"}

\item{count}{Ball-strike count. Defaults to all}

\item{startdate}{Start date}

\item{enddate}{End date}

\item{stand}{Batter handedness}

\item{throws}{Pitcher handedness}

\item{size}{Size of the markers}

\item{heatmap}{Defaults to FALSE}

\item{pitchtypes}{Defaults to all}

\item{color.code}{Defaults to "pitchtype"}

\item{save}{Whether to save the graph. Defaults to FALSE}

\item{path}{Where to save the graph. Defaults to the current working directory}
}
\description{
Graph sz
}

